checking in `jquery.dropdown.css`
since the repository is often used as a distribution package, in
addition to development environments. This allows better flexibility
for consuming the library.

Note that future changes to `jquery.dropdown.less` will require
contributors to also re-compile `jquery.dropdown.css`, but this
shouldn't be a problem since they're already required to re-compile
`jquery.dropdown.min.css`